+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # EDGEGRID GOLANG RELEASE NOTES
 
+## 7.3.0 (September 19, 2023)
+
+#### FEATURES/ENHANCEMENTS:
+
+* ClientLists
+  * Updated `GetClientListResponse` and `UpdateClientListResponse` to include `GroupID`
+
+* GTM
+  * Added custom error `ErrNotFound` that can be used to check if GTM api retuned 404 not found
+
+* HAPI
+  * Added `GetChangeRequest`
+
+* Updated `yaml.v3` dependency
+
 ## 7.2.1 (August 25, 2023)
 
 #### BUG FIXES: